Time: 0 ms (100%), Space: 8.7 MB (44.36%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-double power(double x,int n){
+double power(double x, long long n){
     if(n==0) return 1;
     double half=power(x,n/2);
     if(n%2==0)return half*half;
